Fixed JS error caused by disabling as a response to selection
If you select an item from the dropdown and then disable the component
as a reaction to the event, we would hit a JS error in getInputNode().

This makes it so we always use the existing focus() method when setting focus
and that we check for whether or not we're disabled before attempting
to set focus.