Fix for incorrect placeholder comparison
Placeholder was compared to state.placeholder,
which in the case of a selected value when
multi=false, is actually the label of the
selected item. This means this comparison
was (almost) always returning true, and therefore
always recalculating the values.

This didn't cause any major issue, just
caused extra unnecessary work.

(It also meant that the options being
set later than the value worked, as
the value labels were always recalculated)
meant that