Update to node.js v4.x.x
Updates jsdom to 6.5.x, and fixes API changes Adds a type for
unexpected for a NodeList, as the type is now an object, rather than
an array. This may later be fixed upstream in unexpected-dom.

Also fixes an issue with the link navigation tests. These were relying
on previous tests running first - if the tests run in another order,
they failed.  The initial URL for the click tests is now set as a full
URL, as the new JSDOM (6.x.x) validates it, and errors if an incomplete
URL is used.

Fixes #492