Only filter options when required.
Previously the filter was called in `Select#render`, which is unidiomatic for a React.js component. Instead cache the filtered list of options in `Select#state`, preferring to set this property in `getInitialState`, `componentWillReceiveProps`, and whenever the search term changes (now `setInputValue`).

Also cache `valueArray` rather than calculating it repeatedly.

Closes JedWatson/react-select#719