Don't normalize the input text on creatable
This prevents entered text from being case-folded / having its
diacritics stripped and thus being mangled, by splitting the inputValue
(user's entered value) and normalizedInputValue (the value used for
loadOptions).

Similar to the fix in #1329, except that the behavior is always on (not
controlled by a prop), and the loadOptions behavior is unchanged -- the
loadOptions fn always gets the normalized value, as before.

Fixes #1106.