Update README.md to include 'valueComponent' prop
Thanks to some very helpful feedback I received in this repo (and reading through some issues other users were having) I found there was a valueComponent prop that I could utilize that made my life so much easier.  I think it needs to be included in the list of available props in the docs.