Prefer more descriptive esm word instead of es
es has different meanings. Some call it es5+ syntax or just es6.
There is a precedent of using `es` for untranspiled code in
[materia-ui](https://unpkg.com/@material-ui/core@1.0.0/es/).

In this diff I propose to rename `es` to `esm` as a shorthand for
ES Modules.

`jsnext:main` is deprecated in rollup community (it came from rollup and
dead in it) and shouldn't be used anymore. `module` is enabled by
default like in webpack.