Add a sentence describing what "Creatable" does
I wasted 30 minutes trying understand how to enable users to add new options as well as choosing from existing options. Turns out this capability was there all along in the Creatable component, but it wasn't actually spelled out anywhere in the docs. This PR adds one line to the home page to explain the use of the Creatable component.