Disable use of ResizeObserver for menu auto-update
Resolves https://github.com/JedWatson/react-select/pull/5256#issuecomment-1278916683.

We do not have an official browser support policy, but it would be nice to delay relying on `ResizeObserver` for a major bump and make our official browser support policy clear.

I do not believe we anticipate the `ResizeObserver` getting triggered in many situations anyway since the menu itself shouldn't be resizing. The scroll events are much more critical to auto-updating working properly.