Use ResizeObserver to update menu position if available
Fixes https://github.com/JedWatson/react-select/issues/5403.

I mistakenly thought in https://github.com/JedWatson/react-select/pull/5381 that the resize event was placed on the control element, but's it really just on the scroll parents of the control element.

This PR enables `ResizeObserver` if it's available. The referenced issue demonstrates a good use-case for it.